+++ core-ng/src/test/java/core/framework/internal/redis/RedisSortedSetOperationTest.java
@@ -14,7 +14,7 @@ class RedisSortedSetOperationTest extends AbstractRedisOperationTest {
     @Test
     void add() {
         response(":1\r\n");
-        boolean added = redis.sortedSet().add("key", "v1", 100, false);
+        boolean added = redis.sortedSet().add("key", "v1", 100);
 
         assertThat(added).isTrue();
         assertRequestEquals("*4\r\n$4\r\nZADD\r\n$3\r\nkey\r\n$3\r\n100\r\n$2\r\nv1\r\n");
@@ -23,9 +23,9 @@ class RedisSortedSetOperationTest extends AbstractRedisOperationTest {
     @Test
     void addOnlyIfAbsent() {
         response(":0\r\n");
-        boolean added = redis.sortedSet().add("key", "v1", 100, true);
+        int added = redis.sortedSet().add("key", Map.of("v1", 100L), true);
 
-        assertThat(added).isFalse();
+        assertThat(added).isEqualTo(0);
         assertRequestEquals("*5\r\n$4\r\nZADD\r\n$3\r\nkey\r\n$2\r\nNX\r\n$3\r\n100\r\n$2\r\nv1\r\n");
     }
 
@@ -36,6 +36,15 @@ class RedisSortedSetOperationTest extends AbstractRedisOperationTest {
 
         assertThat(values).containsExactly(entry("v1", 100L), entry("v2", 200L));
         assertRequestEquals("*5\r\n$6\r\nZRANGE\r\n$3\r\nkey\r\n$1\r\n0\r\n$2\r\n-1\r\n$10\r\nWITHSCORES\r\n");
+    }
+
+    @Test
+    void rangeByScore() {
+        response("*4\r\n$2\r\nv1\r\n$3\r\n100\r\n$2\r\nv2\r\n$3\r\n200\r\n");
+        Map<String, Long> values = redis.sortedSet().rangeByScore("key", 100, 200);
+
+        assertThat(values).containsExactly(entry("v1", 100L), entry("v2", 200L));
+        assertRequestEquals("*8\r\n$13\r\nZRANGEBYSCORE\r\n$3\r\nkey\r\n$3\r\n100\r\n$3\r\n200\r\n$10\r\nWITHSCORES\r\n$5\r\nLIMIT\r\n$1\r\n0\r\n$2\r\n-1\r\n");
     }
 
     @Test